Refactor: Move ServerRole to duels-core module
