+++ duels-core/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation project(':duels-api')
+}
+
