+++ duels-core/src/main/java/com/emmerichbrowne/duels/core/ServerRole.java
@@ -1,0 +1,7 @@
+package com.emmerichbrowne.duels.core;
+
+public enum ServerRole {
+	LOBBY,
+	GAME
+}
+
