+++ duels-game-plugin/src/main/java/com/emmerichbrowne/duels/game/DuelsGamePlugin.java
@@ -1,12 +1,11 @@
 package com.emmerichbrowne.duels.game;
 
 import com.emmerichbrowne.duels.DuelsPlugin;
-import com.emmerichbrowne.duels.ServerRole;
 
 public final class DuelsGamePlugin extends DuelsPlugin {
 	@Override
-	public ServerRole getServerRole() {
-		return ServerRole.GAME;
+	public com.emmerichbrowne.duels.core.ServerRole getServerRole() {
+		return com.emmerichbrowne.duels.core.ServerRole.GAME;
 	}
 }
