+++ duels-plugin/src/main/java/com/emmerichbrowne/duels/ServerRole.java
@@ -1,5 +1,6 @@
 package com.emmerichbrowne.duels;
 
+@Deprecated
 public enum ServerRole {
 	LOBBY,
 	GAME
